Refactor TimelineControl to not use SkaaGameDataLib
Initial refactor that uses many less getters, setters, and events. The timeline control is now given a set of frames when the ActiveSprite is changed (this might be a good case for moving that code into a setter for the ActiveSprite). All functionality is seemingly preserved, but I need to come up with a unit testing methodology for C# first (preferably the most dead-simple there is).