Clean up TimelineControl methods for clarity
Updated the method/field names for clarity and changed all the
fields/variables to private. Added regions for organization and commented
out the unused UpdateCurrentFrame() overloads. Also added a
SetCurrentFrameTo(Image) method in order to eliminate the possibility of
accidentally editing the current image as UpdateCurrentFrame() would do.
The updating should be, and now is, handled separately to provide more
control to the developer/UI.