Fix TimelineControl's null handling
When closing a Project, the UI is updated based on whether ActiveSprite
and/or ActiveFrame is null, just like when opening files. TimelineControl
now allows for and properly handles this case with some simple null
checking. I also noticed an erroneous extra line in
RaiseActiveFrameChangedEvent() that set the picBoxFrameImage a second
time. This has been commented out... I'm not 100% sure why it was there
but removing it appears to have not broken anything else.