Add DbfFile saving, make Save() generic, clean up
In Project, consolidate the various Save() methods into a single generic
method. The functionality is now there to save a DataTable as a DbfFile
but it is not exposed in the UI yet. The user needs to be able to specify
which table they're saving.

In SkaaEditorMainForm, cleaned up SaveFile and fixed the SaveFileDialog's
filter. It does not yet call the Project's function to save the DataTable.