Add ObjectListViewContainer test form
Simply shows an ObjectListView, anchored to the form. Call SetDataSource()
to set its DataSource to an IListSource (like DataTable). It currently
works for the SFRAME table in std.set. It needs to be flexible and will
be used to allow the user to select which files they're going to save and
how, when more than one data source is used for ActiveGameSet.