Remove ProjectTypes functionality
Eliminated the entire class and its use throughout. It's been unused for a
while now and will not likely be useful. It was originally to be used for
saving files in a certain format but it would be better to allow the user
to have changes to multiple file types and not be restricted to particular
ProjectType.