Add DataSource to DataSet's Extended Properties
In addition to moving DataSet's extension methods to their own file, there
is a new extension method to add a datasource. This is modified/added to
every time a new DataTable has to be created when loading a file. It's a
list of the data sources (files) that contributed to the current DataSet.
This will allow for tracking all the data sources to ease saving files.