Create ProjectManager and Refactor Project
This is very much in progress. While MVP applies simply to most of the
project with the Model being SkaaGameDataLib, View being the various forms
and controls and Presenter being the help classes, Project is not part of
SkaaGameDataLib. Project is essentially a Model as far as the main forms
are concerned because it saves/loads data. Therefore, ProjectManager can
act as its Presenter.

I say "very much in progress" because, ideally, Project will act as the
Presenter for the entire UI, once I figure out how to structure it.
Project was getting too unweildy with the previous incarnation.

The old Project has been renamed oldProject while I build the new one.