Start Cleanup of DataSetExtensions
This code is inconsistent and was just thrown together. I've started to
clean it up by standardizing the methods and how they're used. I also
added static strings to standardize the ExtendedProperties that are used
to indicate the data source.

DataTable still needs its own extension class to help in this cleanup.