Continue MDI Migration
I've added the ability to load palettes, sprites and the standard game
set. There are likely bugs abound, but the code base is there. Most is
copy-pasted from the old version.

The primary difference now is that ActivePalette is based on the
ActiveSprite for the ActiveDocument in the DockPanel. This is so the
palette updates automagically when the user changes tabs between different
sprites.

I've also begun farming out the old BeginOpenFile() function to
FileDialogExtensions, which will host extension methods for OpenFileDialog
and SaveFileDialog. The idea is the same; to reduce code duplication...
though we're not quite there yet.