Continue MVP Implementation with Generic Loading
There is a new interface, ICustomOpenFileDialog that the presenters will
implement in order to serve up an OpenFileDialog with various preset
options (FileExtension, etc). This allowed for a generic Open<T>() method.

I began implementing INotifyPropertyChanged and events so the UI can be
driven by the Project instead of the other way around. We'll need to use a
better ObservableCollection implementation that can handle PropertyChanged
notifications from its objects.

Also began implementing a better way to provide the property's name when
firing events, removing all the magic strings which just get in the way of
refactoring.