Re-implement FileFormats for Loading Files
A FileFormat property was added to PresenterBase to be used for the Open()
method. The params were also standardized for the various types of Load()
methods that may be called... I still don't like how hackish it seems to
not have strong typing in there. I may implement a LoadArgs class to clear
that up.