Add New Interface and Base Class for Sprite Types
ResIdxMultiBmp and SpriteSpr types are essentially the same thing, except
the former is just a collection of related images instead of an animated
sprite and it contains a DataTable in the ResIdx header. SpriteSpr's data
is stored as part of the SFRAME table in std.set, so it technically
doesn't have its own DataTable, just a DataView, if and when std.set is
loaded. Obviously, this means that the files have different formats to
result in basically the same data.

The new interface, IMultiImagePresenter, and the abstract class,
MultiImagePresenterBase, allow for better generics use, and the reduction
of confusing params (ab)use, since ProjectManager's generic Open method
can now statically check whether the provided type parameters require a
ColorPalette to be set.