Fix Bug Saving SPR Files W/Out Transparent Pixels
The old code was hacked together and a total mess. The new
GetRleBytesFromBitmap method can properly handle creating SPR files when
bitmaps don't have any transparent bytes and/or don't have transparent
bytes as the last several pixels (in the bottom-right of the image).

Cleaned up all the old code and comments and added/clarified comments.