Eliminate DbaseIIIDataColumn for DataColumnExtensions
Using a custom DataColumn type meant that .NET's built-in merging (like
DataSet.Merge()) does not work. Taking a look at the reference source, the
Merge methods use DataColumn as the base type, not any kind of interface
or other abstraction.

This resulted in the loss of the ByteLength field when merging DataSets.
The new DataColumnExtensions class implements an ExtendedProperties
element named "ByteLength" that accomplishes the same thing. This is now
in line with how the other extensions affect the base System.Data types.