Fix Bug With Differing Columns in Sprite DataViews
SkaaSprite was based primarily on SPR files so a lot of the code was
specific to those formats and ResIdxMultiBmp-formatted RES files, like
i_button.res for example, were hacked in around that.

SkaaSprite's MatchFrameOffsets was renamed to MatchFramesToDataView and
will now find matching frames in its Frames list based first off of the
offset and, failing that, off the name. This is because SPR files have no
frame name data in them so this was used to find those frames in std.set.
These frames had their offsets calculated when the file was read. This
allowed for opening an SPR file and opening std.set afterwards.
ResIdxMultiBmp files have frame names in them so that can be used as a
fallback.

Additionally, an SPR's frames' names and offsets are stored in columns
called "FILENAME" and "BITMAPPTR". This is in std.set's SFRAME table and
is also used in other files. However, ResIdx files don't have field names
like DBF files, so they're assigned "FrameName" and "FrameOffset" by the
DataRowExtensions class. I chose to use these rather than BITMAPPTR and
FILENAME primarily because FILENAME would be confusing since there are no
actual files for these frames; this is from the original developers.