Move IMultiImagePresenter & IFrame in SpriteViewer
IMultiImagePresenter allows SpriteView to maintain awareness of a simple
object that has an ActiveFrame, which required IFrame.

Adding an IFrame to SpriteViewer allows SpriteViewer to remain generic
while being easier to deal with than using just indexes of objects. Now
both TimelineView and ListView have an ActiveFrame that updates when an
object is selected in ListView or TimelineView is used to navigate to a
frame.

ColorPalettePresenter was removed from IMultiImagePresenter because it's
not necessary on an interface level and is only implemented in
MultiImagePresenterBase, where the palette is necessary for loading the
Bitmaps. In SpriteViewer, the IFrame objects already have Bitmaps so their
palettes are already set (and not necessary unless loading/saving, which
SpriteViewer doesn't do).

This also allowed eliminating some confusing events in the main form and
required a bunch of null checks throughout. Now, the main form can just
use a single SetActiveSprite() method to handle updating the UI on any
kind of changes without further duplication of various events due to
ActiveFrame changes.