Change RawFormat to Png For Pasting
Since pasting an image would compare against an Image in memory, there is
no encoder set for Image.RawFormat. Using Png preserves transparency and
should work for most purposes, though this hasn't been tested. I'd rather
not test for the existence of an encoder because this needs to be quick.