Improve Clarity, Adjust FileTypes, Add Documentation
No major changes, just a lot of typo fixes and changes for clarity.

The FileTypes enum was updated for clarity, mostly so its values made more
sense when referenced in documentation, which improves Intellisense usage.

The DataSetExtensions class was accidentally left within the SkaaEditorUI
namespace when it was moved. This was changed to SkaaGameDataLib.

ResourceDatabase was renamed to ResourceDefinitionReader since it's
no longer used as an instance type and is now only for reading
definitions. It was also changed to static; both changes are just
oversights
from previous refactoring. Additionally, its MaxRecordCount property was
changed to a private const int and renamed for clarity. It is no longer
needed outside the class so it didn't need to be exposed.