Fix DataSet Merging With ResIdxMultiBmp
Merging was only set up within GameSetPresenter but, since
MDISkaaEditorMainForm was creating a new GameSetPresenter object when
loading a game set, it was useless. This was moved to the UI layer,
eliminating the need for passing the params for merging.

If the game set is already open, the ResIdxMultiBmp's table will always be
merged. If not, ResIdxMultiBmp will create a new DataSet.

If the game set is opened after a ResIdxMultiBmp, then the UI layer will
handle whether or not to merge. It is currently merged by default.