Fix Bug and Add DataTable Extension Method
Re Bug: The old project saved files manually. Now,
DataRowExtensions.WriteResDefinition writes out the offset that is stored
in the row instead of being passed the offset by the calling method.

DataTableExtensions used to only write out a definition for itself when
saving a game set. It will now write ResIdx definitions to the specified
stream for each of its DataRows.