RecalculateFrameOffsets When ActiveFrame Changes
This is needed to properly save ResIdxMultiBmp files and to update the
game set when an SPR file is saved. The offsets are used by the game to
find the bitmaps in a memory buffer so they need to be recalculated any
time a frame is updated. When frames are updated/edited, their pixels
change due to the RLE encoding so every frame in the file following that
frame also has to be updated.