Fix Bug in GetGameSetStream w/ Mis-Counted Records
I'd forgotten to only get a record_count for those tables that are in the
specified set. When a ResIdxMultiBmp was open, it added an additional
table to the set, which is not part of the standard game set.

Also updated a bit of documentation and fixed an accidental use of ==
instead of =! when deciding whether or not to enable the Save Game Set
menu option.

The changes in MDISkaaEditorMainForm.Designer.cs are due to a botched
stash merge, that after checking out a different branch, appeared
impossible to undo. Oops...