Add Method to Update FramePresenters With DataView
Previously, FramePresenter's properties (Name, BitmapOffset, Bitmap) just
returned it's GameObject's equivalent properties. Now, for the sake of
consistency, it maintains its own copies of those values so it needs to be
updated with the new Name and BitmapOffset values from std.set. The
underlying SkaaFrame is already updated in a similar fashion by the
SkaaSprite code.