Remove Non-Idx Reading/Writing (see #41)
The isIdx option was always true except when checking for FileFormats,
where a file that couldn't be identified, would also be checked for
non-ResIdx definitions. It turns out that these don't exist.

The isIdx parameter was removed along with the non-existent Res-based
(8-char) fields in ResourceDefinitionReader.

OffsetSize was renamed to ResIdxOffsetSize.